move conversion of short buffer to float to C++, it's slow af in Java

+++ app/src/main/java/re/bass/beatnik/audio/AudioInput.java
@@ -15,7 +15,7 @@ public interface AudioInput extends Startable {
     int getFrameSize();
 
     interface AudioListener {
-        void onAudio(float[] buffer);
+        void onAudio(short[] buffer);
         void onStart();
     }
 }
